Log switch working.
git-svn-id: svn+ssh://svn.open.ch/pkg-OSAGfsperf/trunk@27 2079a70e-0e9d-4454-99f1-4d4c78da5c83

+++ syncedfs/sighandlers.c
@@ -9,13 +9,32 @@
 
 #include <signal.h>
 #include <unistd.h>
+#include "config.h"
 #include "sighandlers.h"
+#include "../syncedfs-common/lib/error_functions.h"
+#include "log.h"
+#include <stdint.h>
+
+extern sig_atomic_t switchpending;
+extern sig_atomic_t writepending;
+static sig_atomic_t pidpending;
 
 void handleSIGUSR1(int sig, siginfo_t *siginfo, void *ucontext) {
-    // switch log
-    sleep(1);
+    // we don't have to block SIGUSR1, it is done automatically by sigaction
+    // if we can't do the switch right away, store pid for further use
+    if (writepending == 1) {
+        switchpending = 1;
+        pidpending = siginfo->si_pid;
+        return;
+    }
     
+    switchLog();
+    switchpending = 0;
     
-    // when everything is done, send signal back
-    kill(siginfo->si_pid, SIGUSR1);
+    // if we were called directly by operating system, we know pid from siginfo
+    // if we were called by log operation, we used stored pid
+    if (siginfo != NULL)
+        kill(siginfo->si_pid, SIGUSR1);
+    else
+        kill(pidpending, SIGUSR1);
 }
